backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
requests==2.32.3
google-generativeai==0.8.3
Pillow==11.0.0
pytesseract==0.3.13
psycopg[binary,pool]==3.2.3
pgvector==0.3.4
python-multipart==0.0.12
tenacity==9.0.0
debugpy==1.8.5
pydantic==2.4.